Advanced HTML Links
Link, - Adds a link to the page.
Note: target="_blank" opens the link in a new window.

Anchor link-Link to a section on the same page

Mailto link,-The <a> HTML element (or anchor element), with its href attribute, creates a hyperlink to web pages, files, email addresses, locations in the same page, or anything else a URL can address.

Pseudo-classes selector-A CSS pseudo-class is a keyword added to a selector that specifies a special state of the selected element(s). For example, the pseudo-class :hover can be used to select a button when a user's pointer hovers over the button and this selected button can then be styled.